mps issue test

+++ deep_sort/preprocessing.py
@@ -38,7 +38,7 @@ def non_max_suppression(boxes, classes, max_bbox_overlap, scores=None):
     if len(boxes) == 0:
         return []
 
-    boxes = boxes.astype(np.float32)
+    boxes = boxes.astype(np.float)
     pick = []
 
     x1 = boxes[:, 0]
